chore: rename state.rs to cluster_state.rs
